feat: new types
